Finalizando entrega do projeto

+++ new-code/app/controllers/booksController.js
@@ -1,6 +1,8 @@
 const res = require('express/lib/response');
 const Joi = require('joi');
 const BookModel = require('../models/booksModel');
+const jtw = require("jsonwebtoken");
+const config = require("../../config/auth.json");
 
 const Book = require('../models/booksModel');
 
@@ -9,6 +11,9 @@ const emailUsuário2 = "usuario2@gmail.com"
 const emailBibliotecario = "bilbiotecario@gmail.com"
 const passwordUsers= "123"
 
+const generateToken = (properties) => jtw.sign(properties, config.secret, {
+    expiresIn: 3600,
+} )
 
 const schema = Joi.object().keys({
     nome: Joi.string().required(),
@@ -18,11 +23,11 @@ const schema = Joi.object().keys({
 });
 
 module.exports = class Books {
-    static async apiGetAllBooks(req, res, next) {
+    static async apiGetAllBooks(req, res) {
         console.log('Controller book - get books');
         try {
             const books = await Book.getAllBooks();
-            if (!books) {
+            if (books.length<=0) {
                 res.status(400).json('Não existe um livro cadastrado');
                 return;
             }
@@ -38,20 +43,18 @@ module.exports = class Books {
             const booksId = req.params.id;
             const book = await BookModel.getBook(booksId);
             if (!book)
-                return res.status(404).json(`Não existe book cadastrada com o id ${booksId}.`);
+                return res.status(404).json(`Não existe livro cadastrada com o id ${booksId}.`);
             else
                 return res.status(200).json(book);
         } catch (error) {
             console.log(`[Controller - get book by id error] ${error}`);
             res.status(500).json({ error: error });
         }
-
     }
 
-    static async addBook(req, res, next) {
+    static async addBook(req, res) {
         console.log(`[Add Book Controller]`, req.body);
         const { error, value } = schema.validate(req.body);
-        // console.log(`[Controller add Book erro: ] ${value} - ${error.details}`);
         if (error) {
             const result = {
                 msg: `Livro não incluído. Campos não foram preenchidos corretamente`,
@@ -61,8 +64,9 @@ module.exports = class Books {
             return;
         }
         try {
-            const addedBook = await Book.addBook(req.body);
-            res.status(200).json(addedBook);
+            if(req.id!=emailBibliotecario) return res.status(403).json(`Pessoa não autorizada`)
+            const addedBook = await Book.addBook(req.body, req.id);
+            res.status(200).json(`Livro ${req.body.nome} adicionado com sucesso`);
         } catch (error) {
             res.status(500).json({ error: error });
         }
@@ -70,6 +74,7 @@ module.exports = class Books {
     
     static async deleteBook(req, res) {
         try {
+            if(req.id!= emailBibliotecario) return res.status(403).json(`Pessoa não autorizada`)
             const deleteBook = await BookModel.deleteBook(req.params.id);
             if (!deleteBook) return res.status(404).json(`Não existe livro cadastrado com o id ${req.params.id}.`);
             return res.status(200).json(`Livro ${req.params.id} removido com sucesso`);
@@ -80,7 +85,11 @@ module.exports = class Books {
 
     static async loanRequisityBook(req, res){
         try {
-            
+            console.log(req.id);
+            if(req.id==emailBibliotecario) return res.status(403).json(`Bibliotecários não podem fazer requisição de empréstimo`)
+            const book = await BookModel.loanRequisityBook(req.params.id, req.id);
+            if(!book) return res.status(404).json(`Livro não encontrado`);
+            res.status(200).json(`Livro ${req.params.id} requisitado com sucesso`);
         } catch (error) {
             res.status(500).json({ error: error });
         }
@@ -88,20 +97,37 @@ module.exports = class Books {
 
     static async loanBook(req, res){
         try {
-            
+            if(req.id== emailBibliotecario) return res.status(403).json(`Bibliotecários não podem fazer empréstimo de livro`)
+            const obj = {};
+            obj["emprestado"] = true;
+            obj["emprestadoPara"] = req.id;
+            const book = await BookModel.loanBook(req.params.id, obj);
+            if(!book) return res.status(404).json(`Livro não encontrado`);
+            res.status(200).json(`Livro ${req.params.id} emprestado com sucesso`);
         } catch (error) {
-            
+            res.status(500).json({ error: error });
         }
     }
 
     static async devolutionBook(req, res){
         try {
-            //Criar uma pai de delete para complementar a api de update, criar schema para validacao
             const book = await Book.devolutionBook(req.params.id);
             if(!book) return res.status(404).json(`Livro ${req.params.id} não encontrado`);
             res.status(200).json(`Livro ${req.params.id} devolvido com sucesso`);
         } catch (error) {
             res.status(500).json({ error: error });         
+        }
+    }
+
+    static async removeRequisityBook(req, res){
+        try {
+            const book = await BookModel.getBook(req.params.id);
+            if(!book.requisitadoPor.includes(String(req.id))) return res.status(302).json(`Você já removeu seu requisição desse livro`)
+            const removeSignBook = await BookModel.removeRequisityBook(req.params.id, req.id);
+            if(!removeSignBook) return res.status(404).json(`Livro ${req.params.id} não encontrado`);
+            res.status(200).json(`Requisição removida com sucesso`);
+        } catch (error) {
+          res.status(500).json({ error: error });
         }
     }
 
@@ -124,5 +150,5 @@ module.exports = class Books {
         } catch (error) {
           res.status(500).json({ error: error });
         }
-      }
+    }
 }
